// apps/mobile/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
// ...
      records: {
        Row: {
          category_id: number | null
          created_at: string
          date: string
          description: string | null
          duration: number | null
          end_time: string | null
          id: number
          is_user_category: boolean
          start_time: string | null
          user_category_id: number | null
          user_id: string
        }
        Insert: {
          category_id?: number | null
          created_at?: string
          date: string
          description?: string | null
          duration?: number | null
          end_time?: string | null
          id?: number
          is_user_category?: boolean
          start_time?: string | null
          user_category_id?: number | null
          user_id?: string
        }
        Update: {
          category_id?: number | null
          created_at?: string
          date?: string
          description?: string | null
          duration?: number | null
          end_time?: string | null
          id?: number
          is_user_category?: boolean
          start_time?: string | null
          user_category_id?: number | null
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "records_category_id_fkey"
            columns: ["category_id"]
            isOneToOne: false
            referencedRelation: "categories"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "records_user_category_id_fkey"
            columns: ["user_category_id"]
            isOneToOne: false
            referencedRelation: "user_categories"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}